[GR-1111] cfMeDIP report  (#184)
Adds the ability to specify a list of y-axis functions so multiple points can be graphed per x-axis value. This should not impact the definition nor performance of the other graphs. 
Will get back to seeing whether 2 for loops is better or worse than testing for highlight information every iteration