[GR-1205] Home Button on Navbar (#239)
Add a Home button to the navbar in the reports.
Change '.container' to static rather than relative, to fix bug with the Home button jumping down the page. I do not know why this works, or if this will have any consequences. If the navbar is acting weird, this is probably the commit you want to look at