A kludge for crosscheckfingerprints because qc-etl couldn't do its job (#418)
Usually qc-etl keeps two versions around to make migratio invisible to
Dashi. With crosscheckfingerprints, that wasn't possible because the
long cache times are causing Shesmu instability. Having two versions of
caches in there would have probably made the instability a lot worse.
This means that Dashi has to deal with picking the right version. This
is temporary. Once the caches are all switched to the new version, this
can be removed.